requirements.txt
notion-client>=2.0.0
gspread>=5.0.0
google-auth>=2.0.0
slack-sdk>=3.0.0
python-dotenv>=1.0.0
httpx>=0.24.0
certifi>=2023.0.0
